Add files via upload
Virtual Reality is an experimental and innovative framework containing advanced tools and options for creating virtual world applications. Virtual Reality is not a virtual world, nor a stand-alone application, it is a scalable and customizable platform containing some basic modules and based on some fundamental innovative pillars (peer-to-peer architecture, secure communication infrastructure, leg al framework, powerful scripting language); additional modules can be built on top of it on demand.